chore: move price impact color into a hook for max clean beans
