+++ src/lib/utils/index.ts
@@ -17,7 +17,6 @@ import intersection from 'lodash/intersection'
 import isUndefined from 'lodash/isUndefined'
 import union from 'lodash/union'
 import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
-import { bnOrZero } from 'lib/bignumber/bignumber'
 
 export const isKeepKeyHDWallet = (wallet: HDWallet): wallet is KeepKeyHDWallet => {
   return wallet.getVendor() === 'KeepKey'
@@ -261,34 +260,4 @@ export const assertGetChainAdapter = (
   }
 
   return adapter
-}
-
-export const ALLOWED_PRICE_IMPACT_PERCENTAGE_LOW: string = '1' // 1%
-export const ALLOWED_PRICE_IMPACT_PERCENTAGE_MEDIUM: string = '3' // 3%
-export const ALLOWED_PRICE_IMPACT_PERCENTAGE_HIGH: string = '5' // 5%
-export const ALLOWED_PRICE_IMPACT_PERCENTAGE_EXPERT: string = '15' // 15%
-
-const IMPACT_TIERS = [
-  ALLOWED_PRICE_IMPACT_PERCENTAGE_EXPERT,
-  ALLOWED_PRICE_IMPACT_PERCENTAGE_HIGH,
-  ALLOWED_PRICE_IMPACT_PERCENTAGE_MEDIUM,
-  ALLOWED_PRICE_IMPACT_PERCENTAGE_LOW,
-]
-
-type WarningSeverity = 0 | 1 | 2 | 3 | 4
-
-export function warningSeverity(priceImpactPercentage: string | undefined): WarningSeverity {
-  if (!priceImpactPercentage) return 0
-  // This function is used to calculate the Severity level for % changes in USD value and Price Impact.
-  // Price Impact is always an absolute value (conceptually always negative, but represented in code with a positive value)
-  // The USD value change can be positive or negative, and it follows the same standard as Price Impact (positive value is the typical case of a loss due to slippage).
-  // We don't want to return a warning level for a favorable/profitable change, so when the USD value change is negative we return 0.
-  // TODO (WEB-1833): Disambiguate Price Impact and USD value change, and flip the sign of USD Value change.
-  if (bnOrZero(priceImpactPercentage).isLessThan(0)) return 0
-  let impact: WarningSeverity = IMPACT_TIERS.length as WarningSeverity
-  for (const impactLevel of IMPACT_TIERS) {
-    if (bnOrZero(impactLevel).isLessThan(priceImpactPercentage)) return impact
-    impact--
-  }
-  return 0
 }
